Small update
Added FuzzyWuzzy library to help find the best match for a song's name.

Added play functionality and the ability to join/move voice channels.

Bot will now send a comment while searching for a song.

Added options for the audio stream.

Added logic for the 'entry' to be compared against results and find the best match.

Added logic for audio to be reproduced through FFmpeg.