Quick changes
Restructured folders and files: Created a "bot" folder containing "bot.py" (main) and "config.py" (previously "Spotify.py" which was not entirely accurate since I want to not only play Spotify tracks, but Youtube as well).

Changed message sent when a song is not found.

Added an option in ydl_opts: mp3.